# Add AArch64 sha1 and sha2 extension support