--- a/sha1/src/lib.rs
+++ b/sha1/src/lib.rs
@@ -9,8 +9,8 @@
 //! [1]: https://en.wikipedia.org/wiki/SHA-1
 
 #![no_std]
-#[cfg(not(any(target_arch = "x86_64", target_arch = "x86")))]
-compile_error!("crate can only be used on x86 and x86-64 architectures");
+#[cfg(not(any(target_arch = "x86_64", target_arch = "x86", target_arch = "aarch64")))]
+compile_error!("crate can only be used on x86, x86_64 and AArch64 architectures");
 
 #[link(name="sha1", kind="static")]
 extern "C" {