@tarcieri 2020-01-03
Curious if you could wire up Travis's aarch64 support so we could CI this