@linkmauve 2020-01-03
Do you know which flags their CPUs have?  I’ll have a look at doing this CI thing after I’m done with sha256.