@tarcieri 2020-01-03
@linkmauve doesn't look like a224dd2 did the trick... perhaps move the `arch` key to the toplevel?