@tarcieri 2020-01-03
@linkmauve ok, unrelated build failure this time but at least it attempted an ARM build.

That said, it seems like you'll probably need to go back to explicit build matrix entries for `arm64`, but run tests specifically for `sha1` and `sha2`. Something like:

```
cargo test --package sha1 --package sha2
```