@tarcieri 2020-01-03
Whoops, my bad, `--package sha-1`