@linkmauve 2020-01-03
Hmm, it seems like that’s not the correct way to do it. And there is only one build now instead of three.