@tarcieri 2020-01-03
Whoops sorry, those are still the wrong package names. Try:

```
cargo test --package sha1-asm --package sha2-asm
```

That passes locally for me.