@linkmauve 2020-01-03
How do you test locally btw?