@tarcieri 2020-01-03
I was just making sure the command worked. It didn't actually test on ARM (and it seems there aren't specifically tests for these crates, which is unfortunate)

I can try to add them retroactively if we can get the build working.