@tarcieri 2020-01-03
Nice! Let me do another pass on this and I can merge it.

Not sure if I have access to release the associated crates though, unfortunately.