@tarcieri 2020-01-03
My main worry with merging this as-is is the lack of test vectors. You can probably reuse the ones from the `sha-1` and `sha2` crates.