@linkmauve 2020-01-03
I’ve already tested it with those, it was how I found most of the bugs while writing this code. ^^

See https://github.com/RustCrypto/hashes/pull/96