@tarcieri 2020-01-03
That's good, but it'd also be good if they ran in CI too.