@linkmauve 2020-01-03
This would add a test dependency on digest, do you really want that?  Or maybe I could just copy the implementations, but meh.