@tarcieri 2020-01-03
> This would add a test dependency on digest, do you really want that?

I can merge this as/is and we can circle back on that.